+++ services/web/server/src/simcore_service_webserver/catalog/_controller_rest_exceptions.py
@@ -6,7 +6,6 @@ from servicelib.rabbitmq.rpc_interfaces.catalog.errors import (
     CatalogItemNotFoundError,
 )
 
-from ..errors import WebServerBaseError
 from ..exception_handling import (
     ExceptionToHttpErrorMap,
     HttpErrorInfo,
@@ -14,29 +13,7 @@ from ..exception_handling import (
     to_exceptions_handlers_map,
 )
 from ..resource_usage.errors import DefaultPricingPlanNotFoundError
-
-
-class BaseCatalogError(WebServerBaseError):
-    msg_template = "Unexpected error occured in catalog submodule"
-
-    def __init__(self, msg=None, **ctx):
-        super().__init__(**ctx)
-        if msg:
-            self.msg_template = msg
-
-    def debug_message(self):
-        # Override in subclass
-        return f"{self.code}: {self}"
-
-
-class DefaultPricingUnitForServiceNotFoundError(BaseCatalogError):
-    msg_template = "Default pricing unit not found for service key '{service_key}' and version '{service_version}'"
-
-    def __init__(self, *, service_key: str, service_version: str, **ctxs):
-        super().__init__(**ctxs)
-        self.service_key = service_key
-        self.service_version = service_version
-
+from .errors import DefaultPricingUnitForServiceNotFoundError
 
 # mypy: disable-error-code=truthy-function
 assert CatalogForbiddenError  # nosec
