+++ services/web/server/src/simcore_service_webserver/catalog/_service.py
@@ -28,12 +28,12 @@ from servicelib.rest_constants import RESPONSE_MODEL_POLICY
 
 from ..rabbitmq import get_rabbitmq_rpc_client
 from . import _catalog_rest_client_service
-from ._units_service import can_connect, replace_service_input_outputs
-from .controller_rest_schemas import (
+from ._controller_rest_schemas import (
     CatalogRequestContext,
     ServiceInputGetFactory,
     ServiceOutputGetFactory,
 )
+from ._units_service import can_connect, replace_service_input_outputs
 
 _logger = logging.getLogger(__name__)
