+++ services/web/server/src/simcore_service_webserver/catalog/plugin.py
@@ -6,7 +6,7 @@ from aiohttp import web
 from pint import UnitRegistry
 from servicelib.aiohttp.application_setup import ModuleCategory, app_module_setup
 
-from . import _rest_controller, _rest_tags_controller
+from . import _controller_rest, _tags_controller_rest
 
 _logger = logging.getLogger(__name__)
 
@@ -22,11 +22,11 @@ def setup_catalog(app: web.Application):
     # ensures routes are names that corresponds to function names
     assert all(  # nosec
         route_def.kwargs["name"] == route_def.handler.__name__  # type: ignore[attr-defined] # route_def is a RouteDef not an Abstract
-        for route_def in _rest_controller.routes
+        for route_def in _controller_rest.routes
     )
 
-    app.add_routes(_rest_controller.routes)
-    app.add_routes(_rest_tags_controller.routes)
+    app.add_routes(_controller_rest.routes)
+    app.add_routes(_tags_controller_rest.routes)
 
     # prepares units registry
     app[UnitRegistry.__name__] = UnitRegistry()
